Added WDC3 writing (Offsets are not calculated by the lib yet!)

+++ LibWarDB2/Structures/FieldStructure.cs
@@ -5,10 +5,10 @@
         /// <summary>
         /// Size in bits as calculated by: byteSize = (32 - size) / 8; this value can be negative to indicate field sizes larger than 32-bits
         /// </summary>
-        public int Size { get; set; }
+        public short Size { get; set; }
         /// <summary>
         /// Position of the field within the record, relative to the start of the record
         /// </summary>
-        public uint Position { get; set; }
+        public ushort Position { get; set; }
     }
 }
